requirements.txt
# LLM API clients
litellm>=1.0.0
openai>=1.0.0

# Data download
gdown>=4.7.1

# Data processing
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0

# Logging and utilities
python-dotenv>=0.19.0
tqdm>=4.64.0
tenacity>=8.0.0

# Tokenization
tiktoken>=0.1.0

# Google API
google-genai>=1.26.0

# E2B API
e2b-code-interpreter>=1.5.2

# SmolAgents
smolagents>=1.20.0

# Bio
biopython>=1.85.0
playwright>=1.55.0

# Image processing
pillow>=10.0.0

# Google Drive Public File Downloader
gdown>=5.2.0
diskcache>=5.6.3
together>=1.5.25
pydantic>=2.11.7

# HTTP requests
requests>=2.31.0
anthropic>=0.64.0

# Wikipedia
wikipedia>=1.4.0

# platformdirs
platformdirs>=4.5.0

#backend
fastapi>=0.124.0